# api/openapi.yaml
openapi: 3.0.3
info:
  title: Mie & Beurre Bakery API
  description: Backend API for the Mie & Beurre bakery marketplace application.
  version: 1.0.0
  contact:
    name: Mie & Beurre Team
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.mieetbeurre.com
    description: Production
...
security:
  - BearerAuth: []
paths:
...
  /api/bakeries:
    get:
      tags: [Bakeries]
      summary: List bakeries
      description: Returns a paginated list of bakeries with today's schedule. Supports geo-sorting.
      security: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: lat
          in: query
          description: User latitude for distance sorting
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
        - name: lng
          in: query
          description: User longitude for distance sorting
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
        - name: radius
          in: query
          description: Maximum radius in km
          schema:
            type: number
            format: double
      responses:
        '200':
          description: Paginated list of bakery cards
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BakeryCardListResponse'
  /api/bakeries/{id}:
    get:
      tags: [Bakeries]
      summary: Get bakery details
      description: Returns full details for a single bakery.
      security: []
      parameters:
        - $ref: '#/components/parameters/BakeryId'
      responses:
        '200':
          description: Bakery details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bakery'
        '404':
          description: Bakery not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/bakeries/{id}/menu:
    get:
      tags: [Bakeries]
      summary: Get bakery menu
      description: Returns products grouped by category for the given bakery.
      security: []
      parameters:
        - $ref: '#/components/parameters/BakeryId'
      responses:
        '200':
          description: Menu grouped by category
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    $ref: '#/components/schemas/Product'
        '404':
          description: Bakery not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/products/search:
    get:
      tags: [Bakeries]
      summary: Search products
      description: Search products by name across all bakeries with optional filters.
      security: []
      parameters:
        - name: q
          in: query
          description: Search query (product name)
          schema:
            type: string
        - name: category
          in: query
          description: Filter by category
          schema:
            type: string
        - name: excludeAllergens
          in: query
          description: Comma-separated list of allergens to exclude
          schema:
            type: string
        - name: minHealthScore
          in: query
          description: Minimum health score (1-5)
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
      responses:
        '200':
          description: Paginated search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductSearchListResponse'

  # ─── Orders ─────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    BakeryId:
      name: id
      in: path
      required: true
      description: Bakery ID
      schema:
        type: string
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: INVALID_REQUEST
        message:
          type: string
          example: "invalid request body"
...
    Bakery:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        photoUrl:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        ratingAvg:
          type: number
          format: double
          nullable: true
        ratingCount:
          type: integer
        schedule:
          type: array
          items:
            $ref: '#/components/schemas/DaySchedule'
    DaySchedule:
      type: object
      properties:
        day:
          type: string
          enum: [monday, tuesday, wednesday, thursday, friday, saturday, sunday]
        isOpen:
          type: boolean
        openTime:
          type: string
          example: "07:00"
        closeTime:
          type: string
          example: "19:00"
    BakeryCard:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        photoUrl:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        ratingAvg:
          type: number
          format: double
          nullable: true
        ratingCount:
          type: integer
        todaySchedule:
          $ref: '#/components/schemas/TodaySchedule'
        distance:
          type: number
          format: double
          nullable: true
          description: Distance in km (only when geo params provided)
    TodaySchedule:
      type: object
      properties:
        openTime:
          type: string
          example: "07:00"
        closeTime:
          type: string
          example: "19:00"
        isOpen:
          type: boolean
    BakeryCardListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BakeryCard'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
    Product:
      type: object
      properties:
        id:
          type: string
        bakeryId:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        price:
          type: integer
          format: int64
          description: Price in cents
        photoUrl:
          type: string
        allergens:
          type: array
          items:
            type: string
        healthScore:
          type: integer
        available:
          type: boolean
    ProductSearchResult:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
        price:
          type: integer
          format: int64
        photoUrl:
          type: string
        bakeryId:
          type: string
        bakeryName:
          type: string
        allergens:
          type: array
          items:
            type: string
        healthScore:
          type: integer
    ProductSearchListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductSearchResult'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer

    # ─── Orders ─────────────────────────────────────────────────────────────
...